+++ lib/rails_git_hooks/checks.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'checks/base'
+require_relative 'checks/pre_commit'
+require_relative 'checks/commit_msg'
+require_relative 'checks/pre_push'