+++ lib/rails_git_hooks/checks/commit_msg.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module GitHooks
+  module Checks
+    module CommitMsg
+    end
+  end
+end
+
+require_relative 'commit_msg/jira_prefix'