+++ lib/rails_git_hooks/checks/pre_push.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module GitHooks
+  module Checks
+    module PrePush
+    end
+  end
+end
+
+require_relative 'pre_push/run_tests'