+++ lib/rails_git_hooks/core/error.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module GitHooks
-  VERSION = '0.7.1'
+  class Error < StandardError; end
 end